+++ src/GRA.Controllers/ViewModel/MissionControl/Mail/MailDetailViewModel.cs
@@ -1,16 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
 namespace GRA.Controllers.ViewModel.MissionControl.Mail
 {
     public class MailDetailViewModel
     {
         public GRA.Domain.Model.Mail Mail { get; set; }
+        public List<GRA.Domain.Model.Mail> MailThread { get; set; }
+        public GRA.Domain.Model.User User { get; set; }
         public string SentMessage { get; set; }
-        public string ParticipantLink { get; set; }
-        public string ParticipantName { get; set; }
         public bool CanDelete { get; set; }
         public bool CanMail { get; set; }
     }